bash-completions-getter: Properly handle external commands and arguments
In case the completion is using command via "complete -C", we need to
act differently to when we call a function.

In fact we must export some variables so that the external command can use them
(an example is the "aws" command), other than passing the proper
arguments.

Use getops to parse the command line, as it's easier to maintain
compared to the regex parsing (and probably more efficient too).

This will open the road to support more types of completions as the
simplest one provided to bash via -W or -X.