Simplify and velocize the check for already completions
The pattern we were using to filter the completions was causing wrong
results causing the keys/values to be inverted at times and so us to
replace native ksh completions with bash ones.

This has been fixed in the previous commit, however, as per that it's
clear that we can use a simpler check without using any further support
array.

This improve loading the plugin, avoiding uneeded checks and data
duplication.