bash: Honor cursor position when setting COMP_POINT
Bash expects $COMP_POINT to be set to the cursor position, so that it
can complete the word under cursor.

In the wrapper script we were always setting this to the end of the
string, so completing something like:

  > git clon -
  #         ^ cursor here

Was not working. Passing to the bash the cursor position and using it as
COMP_POINT fixes the issue and allows to properly handle all the partial
completions.

Related to https://github.com/romkatv/zsh4humans/issues/27