Reuse as much as possible ZSH completion variables and pass them to script
Change the way we handle the completion script, passing the arguments as
env variables, using ZHS-provided infos instead of just passing the words as
argument and re-compute the missing infos in the script.
The previous mode was a quicker implementation based on the initial
script, but didn't make much sense here, other than creating quoting
issues (fixable using ${(q)cmd} as well, but not really something we
want).

This also allows to ensure that $COMP_CWORD is properly set to the index
of the word under cursor, instead of setting it to the last one by
default. This is an improvement to the previous commit that was only
adapting the absolute cursor position (COMP_POINT)

Fixes other issues mentioned in [1].

[1] https://github.com/romkatv/zsh4humans/issues/27#issuecomment-633431374