Support all the bash-defined completions, not only the ones in files
While it's a common thing that a completed command provides a completion
file with the basename of the command itself, bash provides also many other
completions by sourcing user files and by calling `complete` for multiple
commands.

So, in order to be really complete, let's ask the bash again for its
completion files, by injecting a fake "complete" function that we use to
record all the completions (and completion functions) available.

This implies that before calling `_completion_loader` to check for a
externally provided completion, we need to just look if we already have a
native completion definition, and only in that case look for one via the
dynamic loader.

This behavior can be controlled using the
$ZSH_BASH_COMPLETIONS_FALLBACK_LOAD_NATIVE_COMPLETIONS variable.