Implement compopt and completion -o (options) support
Handle the bash completion options and reflect them in ZSH, somewhat based
on what bashcompinit does, but with some more features.

We currently support these options:
  nospace, nosort, filenames, default, bashdefault, dirnames, plusdirs

The options are always returned as the first argument of the array we return
from the bash script, in case it will be just an empty line.

As per this the completions are mostly completed and the result basically
the same to the one we would get in bash itself.